BUG 35196403 - Coherence client gets stuck while sending data - fix Solaris compilation issue
[git-p4: depot-paths = "//dev/release.cpp/coherence-cpp-v14.1.1.0/": change = 99538]

+++ src/coherence/component/util/TcpInitiator.cpp
@@ -101,7 +101,7 @@ void TcpInitiator::configureSocket(Socket::Handle hSocket) const
         TcpUtil::setSoLinger(hSocket, cSecs);
 
         // wait a maximum of request timeout for a message send to complete
-        hSocket->setSendTimeout(getRequestTimeout());
+        hSocket->setSendTimeout((int32_t) getRequestTimeout());
         }
     catch (Exception::View e)
         {
